requirements.txt
django==3.2.9
djangorestframework
djangorestframework-simplejwt
django-rest-swagger
django-allauth
